pyproject.toml
[project]
name = "gce"
version = "0.0.3.dev0"
description = "Galatea config editor"
requires-python = ">= 3.12"
maintainers = [
    { name="Henry Borchers", email="hborcher@illinois.edu" }
]
authors = [
    {name="University Library at The University of Illinois at Urbana Champaign: Preservation Services", email="prescons@library.illinois.edu"}
]
license = "NCSA"
license-files = ["LICENSE"]
readme = { file= "README.md", content-type="text/markdown" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Natural Language :: English",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: System :: Archiving"

]
dependencies = [
    "pygments>=2.19.2",
    "pyside6-essentials>=6.9.2",
    "galatea>=0.4.2",
    "tomli-w>=1.2.0",
]

[tool.setuptools]
packages = ["gce"]
zip-safe = true
package-dir = {"" = "src"}

[project.gui-scripts]
gce = "gce.__main__:main"

[build-system]
requires = ["setuptools>=80.9.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
test = [
    "coverage>=7.10.7",
    "pytest>=7.4.0",
    "pytest-qt>=4.1.3",
]
dev = [
    "lxml>=6.0.1",
    "mypy>=1.18.1",
    "ruff>=0.13.1",
    "types-pygments>=2.19.0.20250809",
    {include-group = "test"},
]
ci = [{include-group = "dev"}]
freeze = [
    "cmake>=4.1.0",
    "jinja2>=3.1.6",
    "pyinstaller>=6.16.0",
]

[tool.mypy]
mypy_path = "src"

[tool.ruff]
line-length = 79

[tool.uv.sources]
galatea = {index = "uiuc_prescon_python"}

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true

[[tool.uv.index]]
name = "uiuc_prescon_python"
url = "https://nexus.library.illinois.edu/repository/uiuc_prescon_python/simple"
explicit = true
